init
The goal of this branch is to make the amp scoring functionality such that when the driver presses a button the hand off command is triggered while the swerve paths to a premade path that moves perfectly inline with the amp. When the same button is released the note should be shot out into the amp.