Major Changes
OK so here's the logic:
Boolean of if note is in the intake is only ever set in the stored state. The Stored state command ends with a repeat command to run the intake until the note beam is broken.

Then the amp command takes into account if the note is in the intake

The intake command also take into account a boolean (controlled by a binding) weather the command should end with the beambreak or the limitswitch